+++ bandcamp_scraper/email_scraper.py
@@ -22,25 +22,27 @@ class EmailScraper:
     def email_extractor(self):
         for mail in self.mailbox.mailbox:
             if "release" in mail["Subject"]:
-                input = mail.get_body(preferencelist=('html'))
-                input = self.email_encoder(input)
-                if url := self.html_extractor(input):
+                email_body = str(mail.get_body(preferencelist=('html')))
+
+                if "base64" in email_body or "iso-8859" in email_body:
+                    email_body = self.email_decoder(email_body)
+
+                if url := self.html_extractor(email_body):
                     self.database.append({"Count": 1, "Date": mail["Date"], "Url": url[0], "Flag": "0"})
     
-    def email_encoder(self, input):
-        if "base64" in str(input):
-            input_stripped = search("^.+base64(.+)$", str(input).replace("\n", "")) 
-            input = b64decode(input_stripped.groups(1)[0]).decode('utf-8')
+    @staticmethod
+    def email_decoder(email_body):
+        if "base64" in email_body:
+            stripped_body = search("^.+base64(.+)$", str(email_body).replace("\n", ""))
+            return b64decode(stripped_body.groups(1)[0]).decode('utf-8')
 
-        if "iso-8859-1" in str(input):
-            input_stripped = search("^.+iso-8859-1(.+)$", str(input).replace("\n", ""))
-            input = str(input_stripped.groups(1)[0]).encode("utf-8")
-        
-        return input
+        if "iso-8859-1" in email_body:
+            stripped_body = search("^.+iso-8859-1(.+)$", str(email_body).replace("\n", ""))
+            return str(stripped_body.groups(1)[0]).encode("utf-8")
 
-    def html_extractor(self, input):
-        if "just released" in str(input):
-            if match := search(r"<a href=\"(.+)\">", str(input)) or search(r"<a href=3D\"(.+)\">", str(input)):
+    def html_extractor(self, email_body):
+        if "just released" in str(email_body):
+            if match := search(r"<a href=\"(.+)\">", str(email_body)) or search(r"<a href=3D\"(.+)\">", str(email_body)):
                 return match.groups(1)
             else:
                 return False
